# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RetroVision - Retroreflectivity Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signs:
    get:
      operationId: listSigns
      tags: [signs]
      summary: List road signs and pavement markings
      parameters:
        - name: route_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [traffic_sign, pavement_marking, road_stud, delineator, gantry, shoulder_mounted]
        - name: status
          in: query
          schema:
            type: string
            enum: [compliant, warning, critical, unknown]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of signs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignListResponse"
    post:
      operationId: createSign
      tags: [signs]
      summary: Add a new road sign or marking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sign"
  /signs/{id}:
    get:
      operationId: getSign
      tags: [signs]
      summary: Get sign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sign"
    put:
      operationId: updateSign
      tags: [signs]
      summary: Update sign details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sign"
    delete:
      operationId: deleteSign
      tags: [signs]
      summary: Delete a sign record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /signs/{id}/measurements:
    get:
      operationId: getSignMeasurements
      tags: [signs]
      summary: Get all measurements for a specific sign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Measurements for sign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementListResponse"

  # ROUTES
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List highway inspection routes
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteListResponse"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new inspection route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Get route details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Route details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    put:
      operationId: updateRoute
      tags: [routes]
      summary: Update a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouteBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ALERTS
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List inspection reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate a new inspection report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # AI ANALYSIS
  /upload/analyze:
    post:
      operationId: analyzeImage
      tags: [upload]
      summary: Analyze an image for retroreflectivity using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeImageBody"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
components:
  schemas:
...
    Measurement:
      type: object
      properties:
        id:
          type: string
        sign_id:
          type: string
        route_id:
          type: string
        value:
          type: number
          description: "Retroreflectivity value in mcd/lx/m²"
        condition:
          type: string
          enum: [day_dry, day_wet, night_dry, night_wet, foggy, no_street_light, with_street_light]
        method:
          type: string
          enum: [vehicle_mounted, drone, handheld, ai_camera]
        status:
          type: string
          enum: [compliant, warning, critical, unknown]
        irc_standard:
          type: string
          enum: [IRC_67, IRC_35]
        min_required:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
        notes:
          type: string
        recorded_by:
          type: string
        created_at:
          type: string
      required: [id, sign_id, route_id, value, condition, method, status, irc_standard, min_required, created_at]
    MeasurementListResponse:
      type: object
      properties:
        measurements:
          type: array
          items:
            $ref: "#/components/schemas/Measurement"
        total:
          type: integer
      required: [measurements, total]
...
    Sign:
      type: object
      properties:
        id:
          type: string
        route_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [traffic_sign, pavement_marking, road_stud, delineator, gantry, shoulder_mounted]
        location_description:
          type: string
        chainage:
          type: string
          description: "km marker e.g. NH-48 km 342"
        latitude:
          type: number
        longitude:
          type: number
        status:
          type: string
          enum: [compliant, warning, critical, unknown]
        last_measured_at:
          type: string
        last_value:
          type: number
        installed_at:
          type: string
        image_url:
          type: string
        created_at:
          type: string
      required: [id, route_id, name, type, status, created_at]
    SignListResponse:
      type: object
      properties:
        signs:
          type: array
          items:
            $ref: "#/components/schemas/Sign"
        total:
          type: integer
      required: [signs, total]
    CreateSignBody:
      type: object
      properties:
        route_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [traffic_sign, pavement_marking, road_stud, delineator, gantry, shoulder_mounted]
        location_description:
          type: string
        chainage:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
      required: [route_id, name, type]
    UpdateSignBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        location_description:
          type: string
        chainage:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        image_url:
          type: string
    Route:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        highway_number:
          type: string
          description: "e.g. NH-48, NH-8"
        start_chainage:
          type: string
        end_chainage:
          type: string
        length_km:
          type: number
        lane_count:
          type: integer
        type:
          type: string
          enum: [expressway, four_lane, two_lane]
        status:
          type: string
          enum: [active, inactive, under_inspection]
        total_signs:
          type: integer
        compliant_count:
          type: integer
        warning_count:
          type: integer
        critical_count:
          type: integer
        last_inspected_at:
          type: string
        created_at:
          type: string
      required: [id, name, highway_number, type, status, created_at]
    RouteListResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: "#/components/schemas/Route"
        total:
          type: integer
      required: [routes, total]
    CreateRouteBody:
      type: object
      properties:
        name:
          type: string
        highway_number:
          type: string
        start_chainage:
          type: string
        end_chainage:
          type: string
        length_km:
          type: number
        lane_count:
          type: integer
        type:
          type: string
          enum: [expressway, four_lane, two_lane]
      required: [name, highway_number, type]
    UpdateRouteBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        length_km:
          type: number
        lane_count:
          type: integer
...
    Report:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        route_id:
          type: string
        route_name:
          type: string
        type:
          type: string
          enum: [inspection, compliance, maintenance, periodic]
        status:
          type: string
          enum: [draft, generating, ready, failed]
        period_start:
          type: string
        period_end:
          type: string
        total_signs:
          type: integer
        compliance_rate:
          type: number
        critical_findings:
          type: integer
        generated_by:
          type: string
        created_at:
          type: string
      required: [id, title, type, status, created_at]
    ReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
      required: [reports, total]
    CreateReportBody:
      type: object
      properties:
        title:
          type: string
        route_id:
          type: string
        type:
          type: string
          enum: [inspection, compliance, maintenance, periodic]
        period_start:
          type: string
        period_end:
          type: string
        generated_by:
          type: string
      required: [title, type, period_start, period_end]
    AnalyzeImageBody:
      type: object
      properties:
        image_url:
          type: string
        sign_type:
          type: string
        condition:
          type: string
      required: [image_url]
    AnalysisResult:
      type: object
      properties:
        estimated_value:
          type: number
        confidence:
          type: number
        status:
          type: string
          enum: [compliant, warning, critical]
        recommendations:
          type: array
          items:
            type: string
        detected_issues:
          type: array
          items:
            type: string
        analysis_summary:
          type: string
      required: [estimated_value, confidence, status, recommendations, detected_issues, analysis_summary]