# lib/api-zod/src/generated/types/createRouteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RetroVision - Retroreflectivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRouteBodyType } from "./createRouteBodyType";

export interface CreateRouteBody {
  name: string;
  highway_number: string;
  start_chainage?: string;
  end_chainage?: string;
  length_km?: number;
  lane_count?: number;
  type: CreateRouteBodyType;
}