# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RetroVision - Retroreflectivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteStatus } from "./routeStatus";
import type { RouteType } from "./routeType";

export interface Route {
  id: string;
  name: string;
  /** e.g. NH-48, NH-8 */
  highway_number: string;
  start_chainage?: string;
  end_chainage?: string;
  length_km?: number;
  lane_count?: number;
  type: RouteType;
  status: RouteStatus;
  total_signs?: number;
  compliant_count?: number;
  warning_count?: number;
  critical_count?: number;
  last_inspected_at?: string;
  created_at: string;
}