# lib/api-zod/src/generated/types/routeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RetroVision - Retroreflectivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Route } from "./route";

export interface RouteListResponse {
  routes: Route[];
  total: number;
}