# lib/api-zod/src/generated/types/updateRouteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RetroVision - Retroreflectivity Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRouteBody {
  name?: string;
  status?: string;
  length_km?: number;
  lane_count?: number;
}